Fix unnecessary cast in SBM test
Agent-Logs-Url: https://github.com/UofUEpiBio/epiworld/sessions/6abe0295-9525-448a-8e82-f76c62353be8

+++ tests/29a-sbm.cpp
@@ -43,7 +43,7 @@ EPIWORLD_TEST_CASE("SBM expected degree", "[sbm]") {
 
     for (size_t rep = 0u; rep < n_reps; ++rep)
     {
-        model.seed(static_cast<size_t>(rep + 1u));
+        model.seed(rep + 1u);
         model.agents_sbm(block_sizes, mixing_matrix, true);
 
         // Get the edgelist and compute degrees
